Broke the command parser table into two.. One table for directly sent commands (i.e. privmsg), the other for commands sent to a channel.
This saves on having to determine the prefix ('.' for now) and makes it more
configurable. However there's a dirty looking kludge around line 90 of
module.cpp... rather unfortunately :( Only because the core passes us the
command WITH the '.' prefix for now..

It's a step forward, especially if this service will contain a lot of
commands - a modest efficiency thing I suppose.

Or something.