Added a quick fix..
Why?

Let's examine what was happening:

You'd execute a command in a channel:
  .access

Services would be 'smart' and send a message from the target of the message,
but the target was a channel! So the channel can't talk the remote server
would end up ignoring it and trying to KILL it :(

So anyway, yes.. a quick fix :) This is why game's commands aren't working -
something is making services think game is offline... I'm too lazy to bother
with it just now..