Added a few missing bits to the graph, and added some coding to denote optional libraries and which are dynamically loaded shared objects and which are run-time-linked (libraries)..
Also added a few bits which should sort out this graph working with neato.
Neato generates a more compressed graph however it's a little harder to
read. However, since the graph is technically directional (it points down
the chain of linked objects) it's probably better to keep it as a dot formed
graph anyway.