Renamed the conf.* files to config.* - this was inflexibility originally was a product of the autoconf header (config.h) being in the include/exordium directory.
The Config:: class has been turned into a base class for ConfigInternal:: to
hide many unnecessary variables. This included the Modules:: class, of which
has had its header moved into the library's directory since it is only used
internally.

Instead of using the console's hostname for the underling server, two new
configuration variables have been made - underlingHostname and
underlingDescription. Both variables relate to the underling server. If the
hostname is not set, the server will not be created.