Fixed up the DatabaseTable:: class' header template a bit better (although it still needs a bit more work to generate properly written C++ code - the comma on the last enum value is naughty, but my scheme isn't very good)..
I have populated the fieldtypes definition table for autogen with values
already used by the database, and extended it a bit further (such as all the
integer types, all string types etc). I have added translation data for
mysql and postgresql tables, of which I hope the postgresql types are
actually correct.

BTW, the use of 'DATETIME' in mysql didn't seem as portable as 'TIMESTAMP'
(according to my SQL-92 spec docs too).. So I've changed it over to being
TIMESTAMP entirely - this way when the data is handled within exordium it is
consistant, and probably also a lot easier/efficient to rip apart.. So
anyway, be aware of that change.. *sigh* :(