Ahh finally cvs.sf.net responds!
Switched /msg nick set language <code> over to being dynamic - it truely
relies on the languages *loaded* at the calling time now, rather than
depending on a set of hard-coded (and incorrectly spelt, I might add)
language codes, names, handles, whatever..

However, there's a catch. Kine *cannot* possibly begin to provide any level of
findByName() support for a language. Consider all the variations of a language
name possible spanning the languages around the world, and you quickly realise
*why* language codes are preferable. Some languages only have one name, some
have many names referring to the language. Some languages have ambiguous
names, also attached. Some languages have different spellings for their
language names, depending on dialect etc.

But there's a solution to "work around" this problem (yes, it's a
problem, is anyone willing to build a time machine and go back to the dawn
of creation to teach the cave men to spell their language names in a uniform
manner?)

Included is a new feature, where by the request of usage of the language
(that is, '/msg nick set language' with no parameters) will return the
usage information and a list of languages. However, there's a problem which
will crop up one day if Exordium/Kine ever has more than, say, 40 languages
loaded due to protocol limitations on the payload size of PRIVMSG/NOTICE. In
such an event, the string will need to be fragmented.

Also included is the ability to use the default language, so any drooling
moron can forget what language they speak and pick the default one (that is,
of course, unless they forget how to type 'i-default'). This feature is a
suggestion from RFC-2277, and kine's mod_irc2user has used this for some
time now.

Thought may be warranted in the use of setting multiple languages, a'la
mod_irc2user, since it provides people with the ability to select their
languages of preference in the order they choose.

Then again.. that might be a waste of time.