Well... since everyone else has disappeared, I'm still going on with kine. Here I've added some language data which is used by the staff facilities of the irc2 modules within KineIRCd. These are based quite similarly on AustNet's 'helper' status, although instead of using an integer for assigning levels, kine uses a word to define the job of a user. This job name can be used to assist resolving language data, which will appear in the whois information of a staff member if defined correctly.
When exordium is integrated into kine, changing a user's "job" is as simple
as changing the one word job name, as explained in language.def. It's up to
you guys how you wish to use this data, which is exactly why this staff data
is defined in exordium's language files and *not* kine's -- it's completely
dependant on Exordium's implementation.

AustNet defines the following in its example configuration file as 'X' lines:
# Levels 1 to 8 are reserved for ASD
X:1:is able to retrieve passwords, deal with channel suspensions
X:2:is able to retrieve passwords, deal with channel suspensions, global ChanOP access
X:8:is AustNets helper of the week
X:9:is AustNets irc operator of the week
X:200:is an AustNet lamer
X:201:is a pain in da butt
X:203:is a weenor

Okay, well you can see their mentality with the last three. I draw your
attention to the first three, though, where by the job description is
comprised of a few roles.

You fellas may wish to implement clear cut job descriptions, or do 'flag
based' jobs for staff where they can take on many roles as their job at once.
In the latter case, you will need to define the jobs using some sort of
encoding for the job names. Kine defines a limit of 32 characters for the job
name, though, and requires it to be entirely upper-case to avoid confusion
(and simplify run-time language tag ID discovery).

With a flag based system, though, you would need to define many language
tags for each permutation of the encoded job names. The task of creating
such a list would be *easily* simplified with some autogen magic, however.

I personally like the flag-based system for its flexibility, but it does
require more work than simply having a fixed array of job types.

I leave this in your hands, exordium peeps :)