Module start/stop rejig: The Module:: class is in charge of maintaining the pointer to the service, therefore it's now able to be in charge of service registration and deregistration on the network, and also destruction of the service.
Temporarily, the services are being allocated in oblivion -- their
registration will need to potentially co-incide with the underline server
(soon) and be created *on* that server, or otherwise created on the server
kine provides. This is a temporary solution to get the services 'alive'
within kine for now, and is only valid as kine is not currently checking for
the validity of a service's connection in relation to the network.

This involved moving the start/stop functions to realStart() and realStop()
and using start() and stop() as wrappers to these functions, providing the
room for automatic registration/deregistration without the module needing to
put any effort into it.

The start() function in the module now returns the new Exordium::Service*
pointer, or a null (0) if it is unable to start for any reason - this makes
things a bit easier for module developers.. I hope.. :)

I've removed registerService() and serviceQuit(), since these will be
performed inside Module:: soon. This has caused two issues, though -- Firstly
the console is not being registered/deregistered (well, it's not even a service
though.. which does need fixing), and secondly the nick module relied on
registerService() for some odd reason in order to perform the ghosting
function. This also needs to be repaired, but with a fake Kine::User:: object